test(subgraph): refactor utils
