+++ pkg/subgraph/tests/utils/mocked-functions.ts
@@ -1,7 +1,7 @@
 import { Address } from "@graphprotocol/graph-ts";
 import { createMockedFunction } from "matchstick-as";
 
-import { getETHAddress, getETHInt32, getETHString } from "./utils/converters";
+import { getETHAddress, getETHInt32, getETHString } from "./converters";
 
 export function mockedTokenRPCCalls(
   address: string,
